requirements.txt
streamlit
tensorflow-cpu>=2.16.1
numpy>=1.23.5
pillow